restructurize api to apis
