+++ pkg/daemon/api/views/v1/project/helpers.go
@@ -16,31 +16,43 @@
 // from Last.Backend LLC.
 //
 
-package v1
+package project
 
 import (
-	"github.com/lastbackend/lastbackend/pkg/api/views/v1/project"
-	"github.com/lastbackend/lastbackend/pkg/api/views/v1/service"
-	"github.com/lastbackend/lastbackend/pkg/api/views/v1/user"
-	"github.com/lastbackend/lastbackend/pkg/api/types"
+	"encoding/json"
+	"github.com/lastbackend/lastbackend/pkg/apis/types"
 )
 
-func NewUser(obj *types.User) *user.User {
-	return user.New(obj)
+func New(obj *types.Project) *Project {
+	p := new(Project)
+
+	p.User = obj.User
+	p.Name = obj.Name
+	p.Description = obj.Description
+	p.Updated = obj.Updated
+	p.Created = obj.Created
+
+	return p
 }
 
-func NewProject(obj *types.Project) *project.Project {
-	return project.New(obj)
+func (obj *Project) ToJson() ([]byte, error) {
+	return json.Marshal(obj)
 }
 
-func NewProjectList(obj *types.ProjectList) *project.ProjectList {
-	return project.NewList(obj)
+func NewList(obj *types.ProjectList) *ProjectList {
+	p := new(ProjectList)
+	if obj == nil {
+		return nil
+	}
+	for _, v := range *obj {
+		*p = append(*p, *New(&v))
+	}
+	return p
 }
 
-func NewService(obj *types.Service) *service.Service {
-	return service.New(obj)
-}
-
-func NewServiceList(obj *types.ServiceList) *service.ServiceList {
-	return service.NewList(obj)
+func (obj *ProjectList) ToJson() ([]byte, error) {
+	if obj == nil || len(*obj) == 0 {
+		return []byte("[]"), nil
+	}
+	return json.Marshal(obj)
 }
